move CollapsableButtons file
