+++ admin/src/components/HarbourForm.tsx
@@ -39,7 +39,7 @@ import PublishModal from './PublishModal';
 import PublishDetailsSection from './form/PublishDetailsSection';
 import { IonSelectCustomEvent, SelectChangeEventDetail } from '@ionic/core/dist/types/components';
 import CollapsableWrapper from './form/CollapsableWrapper';
-import CollapsableButtons from './CollapsableButtons';
+import CollapsableButtons from './form/CollapsableButtons';
 
 interface FormProps {
   harbour: HarborInput;
